udp fix and crash again

+++ app.cpp
@@ -9,7 +9,7 @@ int main(int argc, char** argv ) {
   game.Initialaze(0, 3);
   int x = 0;
   int y = 0;
-  int value = 0;
+  short int value = 0;
   while (!game.Is_Game_Over()) {
     game.Out();
     std::cin >> x >> y >> value;
